sql: disable ON CONFLICT actions for indexes
After struggling with SQLite legacy ON CONFLICT actions, finally we
decided to remove them. Still, it is allowed to specify action on NULL
conflict and on whole query action (e.g. INSERT OR IGNORE).

This patch also fixes behaviour of INSERT/UPDATE OR REPLACE statement.
Now we are looking at each secondary index to find an entry with given
key in order to delete it (since original space:replace() provides real
replace only in PK). Almost the same happens for UPDATE OR IGNORE: as
far as UPDATE causes implicit deletion of old entry and insertion of
new one, in case of duplicates in secondary indexes we should omit
removal of old entry.

Finally, removed extra cursors allocations from UPDATE and INSERT
processing and moved the rest closer to their real usage.

Closes #3566
Closes #3565
Part of #3561