box: introduce promotion GC
In the previous commit a promotion protocol was introduced. Each
promotion round produces messages stored in the _promotion space.
After a number of promotions the space can become too big. But it
is not necessary to store those promotion rounds after which
another successfull promotion was executed.

This patch introduces garbage collecting, so after each
successfull promotion old rounds are purged.

Follow up #3055