salad: introduce constant-size bit array
In order to implement z-order curve index we need to introduce a
container for bits with some simple operations support.
The main feature of this structure is a constant size - we need
only array the size is a multiple of 64 bits. This allows to avoid
some overhead.
Also it supports bit interleaving with help of preallocated lookup
table.

Some notable things:
  - https://github.com/olegrok/bit_array - contains also tests
  - https://github.com/noporpoise/BitArray